[Tuner] Refactor to separate ROCM-specific code into rocm/ subdirectory
