+++ amdsharktuner/amdsharktuner/rocm/rocm_dispatch_constraints.py
@@ -1,4 +1,4 @@
-# Copyright 2024 Advanced Micro Devices, Inc.
+# Copyright 2026 Advanced Micro Devices, Inc.
 #
 # Licensed under the Apache License v2.0 with LLVM Exceptions.
 # See https://llvm.org/LICENSE.txt for license information.
@@ -8,15 +8,15 @@
 # in the code and runs it.
 
 import math
-import z3  # type: ignore
-from typing import Optional
 from dataclasses import dataclass, field
+from typing import Optional
 
+import z3  # type: ignore
 from iree.compiler import ir  # type: ignore
-
 from iree.compiler.dialects import iree_codegen, iree_gpu  # type: ignore
 
-from . import common
+from .. import common
+from . import rocm_common
 
 
 @dataclass
@@ -76,7 +76,7 @@ def get_mma_intrinsic_constraints(
     acc_layout: MMASingleSubgroupLayout | None = None,
     allow_virtual_mma: bool = False,
 ) -> z3.BoolRef:
-    compatible_intrinsics = common.get_compatible_mma_intrinsics(
+    compatible_intrinsics = rocm_common.get_compatible_mma_intrinsics(
         lhs_type, rhs_type, res_type, mma_intrinsics, allow_virtual_mma
     )
     assert len(compatible_intrinsics) > 0, "No compatible intrinsics found"
@@ -733,7 +733,7 @@ def generate_compilation_infos(
     compilation_infos = []
     for pipeline_options in pipeline_options_list:
         for waves_per_eu in allowed_waves_per_eu:
-            config_dict = common.get_translation_info_config(
+            config_dict = rocm_common.get_translation_info_config(
                 pipeline_options, waves_per_eu
             )
             translation_info = iree_codegen.TranslationInfoAttr.get(
