+++ amdsharktuner/amdsharktuner/tuner_base.py
@@ -1,0 +1,70 @@
+# Copyright 2025 Advanced Micro Devices, Inc.
+#
+# Licensed under the Apache License v2.0 with LLVM Exceptions.
+# See https://llvm.org/LICENSE.txt for license information.
+# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+
+"""Abstract base class for dispatch tuners.
+
+This module is separate from candidate_gen.py to avoid circular imports
+when target-specific tuners (e.g., ROCm tuners) need to inherit from
+DispatchTuner while candidate_gen.py needs to import those tuners.
+"""
+
+from abc import abstractmethod
+from typing import Optional
+
+from iree.compiler import ir  # type: ignore
+
+from . import common, constraint_generator, dispatch_parser
+
+
+class DispatchTuner(dispatch_parser.DispatchParser):
+    """Abstract base class for dispatch tuners.
+
+    Dispatch tuners are responsible for generating tuning configurations
+    for specific types of operations (contractions, convolutions, attention, etc.)
+    targeting specific hardware backends.
+    """
+
+    @classmethod
+    @abstractmethod
+    def supports_root_op(cls, root_op: ir.Operation) -> bool:
+        """Check if this tuner can handle the given root operation."""
+        pass
+
+    @abstractmethod
+    def get_td_spec(
+        self,
+        config_list: list[common.TuningConfiguration],
+    ) -> ir.Module:
+        """
+        Generates a transform dialect spec from a list of TuningConfiguration objects.
+
+        Each TuningConfiguration specifies a name (e.g., "compilation_info") and
+        its corresponding MLIR attribute (e.g., CompilationInfoAttr) to be applied
+        to the dispatch root operation.
+        """
+        pass
+
+    @abstractmethod
+    def get_constraint_generator(self) -> constraint_generator.ConstraintGenerator:
+        """Returns a ConstraintGenerator associated with this dispatch root op."""
+        pass
+
+    @classmethod
+    @abstractmethod
+    def get_dispatch_kind(cls) -> common.DispatchKind:
+        """Returns dispatch kind."""
+        pass
+
+    @abstractmethod
+    def get_knob_assignment(
+        self,
+        config_list: list[common.TuningConfiguration],
+    ) -> Optional[common.KnobAssignment]:
+        """
+        Return a KnobAssignment that records the feature values of a single candidate,
+        retrieved from the `knob_assignment` attribute of its TuningConfiguration.
+        """
+        pass
