+++ amdsharktuner/tests/rocm/__init__.py
@@ -1,0 +1,7 @@
+# Copyright 2025 Advanced Micro Devices, Inc.
+#
+# Licensed under the Apache License v2.0 with LLVM Exceptions.
+# See https://llvm.org/LICENSE.txt for license information.
+# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+
+"""ROCm-specific tests for amdsharktuner."""
